+++ src/dbus_api/api/manager_2_0/props.rs
@@ -1,0 +1,16 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+use dbus::{
+    self,
+    arg::IterAppend,
+    tree::{MTFn, MethodErr, PropInfo},
+};
+
+use crate::{dbus_api::types::TData, stratis::VERSION};
+
+pub fn get_version(i: &mut IterAppend, _p: &PropInfo<MTFn<TData>, TData>) -> Result<(), MethodErr> {
+    i.append(VERSION);
+    Ok(())
+}
