+++ src/dbus_api/filesystem/fetch_properties_2_0/mod.rs
@@ -1,0 +1,2 @@
+pub mod api;
+pub mod methods;
