+++ src/dbus_api/mod.rs
@@ -7,6 +7,7 @@ mod macros;
 
 mod api;
 mod blockdev;
+mod connection;
 mod consts;
 mod event_handler;
 mod filesystem;
@@ -14,4 +15,4 @@ mod pool;
 mod types;
 mod util;
 
-pub use self::{api::DbusConnectionData, event_handler::EventHandler};
+pub use self::{connection::DbusConnectionData, event_handler::EventHandler};
